fix: allow reuse of tx simulation warning banner
