+++ components/brave_wallet_ui/components/extension/confirm-transaction-panel/common/tx_warning_banner.styles.ts
